requirements.txt
faiss_cpu==1.6.3
scipy==1.4.1
matplotlib==3.3.2
numpy==1.19.5
tensorflow==2.5.0
six==1.15.0
scikit_image==0.18.1
faiss==1.5.3
Pillow==8.3.1
skimage==0.0
